pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "loltrack"
version = "0.1.0"
description = "LoL CLI Stat-Tracker: live match tracking and rolling post-game analytics"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "LoL-CLI", email = "noreply@example.com" }
]
dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0",
  "python-dateutil>=2.9",
  "keyring>=24.3",
]

# No CLI entrypoint — web dashboard only

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []
